Moving attendance_marker func to marker.py
